feat: deploy importer as stateless with replicas=2
Relates trustification/trustify#1207

This matches the number of default enabled importers so that they may
be run concurrently.

Ideally, the replicas value would be parameterized, but I'll need some
guidance as it's not clear from the code how best to achieve that.

Downstream request: https://issues.redhat.com/browse/TC-2218

Signed-off-by: Jim Crossley <jim@crossleys.org>